backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
rank-bm25==0.2.2
openai==1.51.2
httpx<0.28
pydantic==2.9.2
